ikev1: Fix prefix length of vendor ID Cisco Unity
Before commit fc0032cee8 ("ike: Add additional Vendor IDs for
third-party implementations") the prefix length of vendor ID
Cisco Unity was hardcoded to 14. Since we need to know the actual
length of this VID to send it, the length can't be overloaded
with a prefix length. Revert part of commit fc0032cee8 to
fix this problem.

Fixes: fc0032cee879 ("ike: Add additional Vendor IDs for third-party implementations")